# backend/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisSchema": {
        "description": "Canonical v1 analysis payload returned with each document version.",
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterAnalysisSchema"
            },
            "title": "Characters",
            "type": "array"
          },
          "status": {
            "enum": [
              "completed",
              "not_requested"
            ],
            "title": "Status",
            "type": "string"
          },
          "turning_points": {
            "items": {
              "discriminator": {
                "mapping": {
                  "identified": "#/components/schemas/IdentifiedTurningPointSchema",
                  "not_applicable": "#/components/schemas/NotApplicableTurningPointSchema"
                },
                "propertyName": "availability"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/IdentifiedTurningPointSchema"
                },
                {
                  "$ref": "#/components/schemas/NotApplicableTurningPointSchema"
                }
              ]
            },
            "title": "Turning Points",
            "type": "array"
          }
        },
        "required": [
          "status",
          "turning_points",
          "characters"
        ],
        "title": "AnalysisSchema",
        "type": "object"
      },
      "ApiErrorCode": {
        "description": "Stable API error codes consumed by the web client.",
        "enum": [
          "VALIDATION_ERROR",
          "EMPTY_FILE",
          "FILE_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "INVALID_ENCODING",
          "MALFORMED_FDX",
          "MALFORMED_PDF",
          "INVALID_NATIVE_DOCUMENT",
          "PROCESS_RESTARTED",
          "UNAUTHENTICATED",
          "AUTHENTICATION_UNAVAILABLE",
          "ACCOUNT_BINDING_CONFLICT",
          "DOCUMENT_NOT_FOUND",
          "TASK_NOT_FOUND",
          "NOT_FOUND",
          "FORBIDDEN",
          "CONFLICT",
          "DOCUMENT_CONTENT_MISSING",
          "CONTRACT_MISMATCH",
          "INTERNAL_ERROR"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "CharacterAnalysisSchema": {
        "description": "Main character analysis stored in the canonical v1 document payload.",
        "properties": {
          "agency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency"
          },
          "choice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choice"
          },
          "external_goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Goal"
          },
          "fear_or_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fear Or Cost"
          },
          "goal_to_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal To Outcome"
          },
          "internal_need": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Need"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "obstacle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obstacle"
          },
          "related_turning_points": {
            "items": {
              "type": "integer"
            },
            "title": "Related Turning Points",
            "type": "array"
          },
          "voice_traits": {
            "default": "",
            "title": "Voice Traits",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CharacterAnalysisSchema",
        "type": "object"
      },
...
      "DialogueSchema": {
        "description": "Dialogue line within a scene.",
        "properties": {
          "character": {
            "title": "Character",
            "type": "string"
          },
          "line": {
            "title": "Line",
            "type": "string"
          }
        },
        "required": [
          "character",
          "line"
        ],
        "title": "DialogueSchema",
        "type": "object"
      },
...
      "EmotionArcPointMappingSchema": {
        "additionalProperties": false,
        "description": "One aggregated emotion-arc point's inclusive source-scene range.",
        "properties": {
          "end_scene_number": {
            "minimum": 1.0,
            "title": "End Scene Number",
            "type": "integer"
          },
          "point_number": {
            "minimum": 1.0,
            "title": "Point Number",
            "type": "integer"
          },
          "representative_scene_number": {
            "minimum": 1.0,
            "title": "Representative Scene Number",
            "type": "integer"
          },
          "start_scene_number": {
            "minimum": 1.0,
            "title": "Start Scene Number",
            "type": "integer"
          }
        },
        "required": [
          "point_number",
          "start_scene_number",
          "end_scene_number",
          "representative_scene_number"
        ],
        "title": "EmotionArcPointMappingSchema",
        "type": "object"
      },
      "EmotionArcSchema": {
        "description": "Valence/tension data (valence 1-7 int, non-appearance 0, tension -3..+3 int).",
        "properties": {
          "characters": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Characters",
            "type": "object"
          },
          "scene_mapping": {
            "items": {
              "$ref": "#/components/schemas/EmotionArcPointMappingSchema"
            },
            "title": "Scene Mapping",
            "type": "array"
          },
          "tension": {
            "items": {
              "type": "integer"
            },
            "title": "Tension",
            "type": "array"
          },
          "valence": {
            "items": {
              "type": "integer"
            },
            "title": "Valence",
            "type": "array"
          },
          "valence_vector": {
            "items": {
              "type": "number"
            },
            "maxItems": 10,
            "minItems": 10,
            "title": "Valence Vector",
            "type": "array"
          }
        },
        "required": [
          "valence",
          "tension",
          "characters",
          "scene_mapping",
          "valence_vector"
        ],
        "title": "EmotionArcSchema",
        "type": "object"
      },
      "IdentifiedTurningPointSchema": {
        "additionalProperties": false,
        "properties": {
          "availability": {
            "const": "identified",
            "default": "identified",
            "title": "Availability",
            "type": "string"
          },
          "change": {
            "minLength": 1,
            "title": "Change",
            "type": "string"
          },
          "involved_characters": {
            "items": {
              "$ref": "#/components/schemas/TurningPointCharacterSchema"
            },
            "minItems": 1,
            "title": "Involved Characters",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "null"
          },
          "scene_number": {
            "minimum": 1.0,
            "title": "Scene Number",
            "type": "integer"
          },
          "tp_number": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Tp Number",
            "type": "integer"
          }
        },
        "required": [
          "tp_number",
          "label",
          "scene_number",
          "change",
          "involved_characters"
        ],
        "title": "IdentifiedTurningPointSchema",
        "type": "object"
      },
      "ImportDraftResponse": {
        "description": "Ephemeral completed-import payload. It is never written to task_events.",
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/AnalysisSchema"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "emotion_arc": {
            "$ref": "#/components/schemas/EmotionArcSchema"
          },
          "expected_version": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Expected Version"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "narrator": {
            "$ref": "#/components/schemas/NarratorSchema"
          },
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneSchema"
            },
            "title": "Scenes",
            "type": "array"
          },
          "source_fountain": {
            "title": "Source Fountain",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version_id": {
            "minimum": 1,
            "title": "Version Id",
            "type": "integer"
          },
          "voice_assignments": {
            "items": {
              "$ref": "#/components/schemas/VoiceAssignmentSchema"
            },
            "title": "Voice Assignments",
            "type": "array"
          }
        },
        "required": [
          "document_id",
          "version_id",
          "title",
          "source_fountain",
          "metadata",
          "scenes",
          "analysis",
          "emotion_arc"
        ],
        "title": "ImportDraftResponse",
        "type": "object"
      },
...
      "NarratorSchema": {
        "description": "Voice configuration for the narrator.",
        "properties": {
          "voice_traits": {
            "default": "",
            "title": "Voice Traits",
            "type": "string"
          }
        },
        "title": "NarratorSchema",
        "type": "object"
      },
...
      "NotApplicableTurningPointSchema": {
        "additionalProperties": false,
        "properties": {
          "availability": {
            "const": "not_applicable",
            "default": "not_applicable",
            "title": "Availability",
            "type": "string"
          },
          "change": {
            "title": "Change",
            "type": "null"
          },
          "involved_characters": {
            "items": {
              "$ref": "#/components/schemas/TurningPointCharacterSchema"
            },
            "maxItems": 0,
            "title": "Involved Characters",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "scene_number": {
            "title": "Scene Number",
            "type": "null"
          },
          "tp_number": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Tp Number",
            "type": "integer"
          }
        },
        "required": [
          "tp_number",
          "label",
          "reason"
        ],
        "title": "NotApplicableTurningPointSchema",
        "type": "object"
      },
      "ReanalysisTaskResultEvent": {
        "description": "Validated emotion-arc draft returned by a completed reanalysis task.",
        "properties": {
          "emotion_arc": {
            "$ref": "#/components/schemas/EmotionArcSchema"
          }
        },
        "required": [
          "emotion_arc"
        ],
        "title": "ReanalysisTaskResultEvent",
        "type": "object"
      },
      "SceneSchema": {
        "description": "Structured scene in a screenplay.",
        "properties": {
          "dialogues": {
            "items": {
              "$ref": "#/components/schemas/DialogueSchema"
            },
            "title": "Dialogues",
            "type": "array"
          },
          "heading": {
            "title": "Heading",
            "type": "string"
          },
          "scene_number": {
            "title": "Scene Number",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "scene_number",
          "heading",
          "text",
          "dialogues"
        ],
        "title": "SceneSchema",
        "type": "object"
      },
...
      "TaskCancelResponse": {
        "properties": {
          "status": {
            "default": "cancel_requested",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "TaskCancelResponse",
        "type": "object"
      },
      "TaskCancelledEvent": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "TaskCancelledEvent",
        "type": "object"
      },
      "TaskCompletedEvent": {
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Document Id"
          },
          "import_draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImportDraftResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "reanalysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReanalysisTaskResultEvent"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Version Id"
          }
        },
        "title": "TaskCompletedEvent",
        "type": "object"
      },
      "TaskErrorEvent": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "TaskErrorEvent",
        "type": "object"
      },
      "TaskProgressEvent": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "percentage": {
            "maximum": 100,
            "minimum": 0,
            "title": "Percentage",
            "type": "integer"
          },
          "phase": {
            "enum": [
              "queued",
              "cancelling",
              "structuring",
              "classifying",
              "reading",
              "building_plot_graph",
              "planning",
              "writing",
              "verifying",
              "analyzing"
            ],
            "title": "Phase",
            "type": "string"
          },
          "received_characters": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Received Characters"
          }
        },
        "required": [
          "phase",
          "percentage",
          "message"
        ],
        "title": "TaskProgressEvent",
        "type": "object"
      },
      "TurningPointCharacterSchema": {
        "additionalProperties": false,
        "description": "One character's goal/conflict/choice/action/change at a turning point.",
        "properties": {
          "action": {
            "minLength": 1,
            "title": "Action",
            "type": "string"
          },
          "change": {
            "minLength": 1,
            "title": "Change",
            "type": "string"
          },
          "choice": {
            "minLength": 1,
            "title": "Choice",
            "type": "string"
          },
          "conflict": {
            "minLength": 1,
            "title": "Conflict",
            "type": "string"
          },
          "goal": {
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "goal",
          "conflict",
          "choice",
          "action",
          "change"
        ],
        "title": "TurningPointCharacterSchema",
        "type": "object"
      },
...
      "VoiceAssignmentSchema": {
        "description": "Assignment of a voice to a speaker (narrator or character).",
        "properties": {
          "speaker": {
            "minLength": 1,
            "title": "Speaker",
            "type": "string"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "voice_traits": {
            "default": "",
            "title": "Voice Traits",
            "type": "string"
          }
        },
        "required": [
          "speaker"
        ],
        "title": "VoiceAssignmentSchema",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Intelligent screenplay analysis and narrative arc workbench API",
    "title": "NARRAVANT API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "cancel_task_api_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Cancel Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/progress": {
      "get": {
        "operationId": "task_progress_api_v1_tasks__task_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TaskProgressEvent"
                    },
                    {
                      "$ref": "#/components/schemas/TaskCompletedEvent"
                    },
                    {
                      "$ref": "#/components/schemas/TaskErrorEvent"
                    },
                    {
                      "$ref": "#/components/schemas/TaskCancelledEvent"
                    }
                  ]
                }
              }
            },
            "description": "Persisted task SSE events; comments may be heartbeat frames."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Task Progress",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}